# openapi/swagger.yaml
components:
  schemas:
...
    internal_httpapi.ErrorResponse:
      properties:
        error:
          type: string
      type: object
    internal_httpapi.FavoritePatchRequest:
      properties:
        favorited:
          type: boolean
      type: object
    internal_httpapi.FavoriteResponse:
      properties:
        favorited:
          type: boolean
      type: object
...
    internal_httpapi.HomeItemsResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/sudoStream_internal_homeshelf.ShelfItem'
          type: array
          uniqueItems: false
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
      type: object
...
    sudoStream_internal_access.LibraryType:
      enum:
      - film
      - series
      - music
      - photos
      - other
      type: string
      x-enum-varnames:
      - LibraryTypeFilm
      - LibraryTypeSeries
      - LibraryTypeMusic
      - LibraryTypePhotos
      - LibraryTypeOther
...
    sudoStream_internal_catalog.Episode:
      properties:
        actions:
          $ref: '#/components/schemas/sudoStream_internal_mediafs.Action'
        episode:
          type: integer
        episodeTitle:
          type: string
        favorited:
          type: boolean
        path:
          type: string
        posterUrl:
          type: string
        season:
          type: integer
        title:
          type: string
        watched:
          type: boolean
      type: object
    sudoStream_internal_catalog.LibraryCatalog:
      properties:
        limit:
          type: integer
        movies:
          items:
            $ref: '#/components/schemas/sudoStream_internal_catalog.Movie'
          type: array
          uniqueItems: false
        offset:
          type: integer
        shows:
          items:
            $ref: '#/components/schemas/sudoStream_internal_catalog.ShowSummary'
          type: array
          uniqueItems: false
        total:
          type: integer
        type:
          $ref: '#/components/schemas/sudoStream_internal_access.LibraryType'
      type: object
    sudoStream_internal_catalog.Movie:
      properties:
        actions:
          $ref: '#/components/schemas/sudoStream_internal_mediafs.Action'
        favorited:
          type: boolean
        path:
          type: string
        posterUrl:
          type: string
        title:
          type: string
        watched:
          type: boolean
        year:
          type: integer
      type: object
    sudoStream_internal_catalog.SeasonEpisodes:
      properties:
        episodes:
          items:
            $ref: '#/components/schemas/sudoStream_internal_catalog.Episode'
          type: array
          uniqueItems: false
        limit:
          type: integer
        offset:
          type: integer
        season:
          type: integer
        total:
          type: integer
      type: object
    sudoStream_internal_catalog.SeasonSummary:
      properties:
        episodeCount:
          type: integer
        season:
          type: integer
      type: object
    sudoStream_internal_catalog.ShowDetail:
      properties:
        actions:
          $ref: '#/components/schemas/sudoStream_internal_mediafs.Action'
        episodeCount:
          type: integer
        name:
          type: string
        posterPath:
          type: string
        posterUrl:
          type: string
        seasonCount:
          type: integer
        seasons:
          items:
            $ref: '#/components/schemas/sudoStream_internal_catalog.SeasonSummary'
          type: array
          uniqueItems: false
        showKey:
          type: string
      type: object
    sudoStream_internal_catalog.ShowSummary:
      properties:
        actions:
          $ref: '#/components/schemas/sudoStream_internal_mediafs.Action'
        episodeCount:
          type: integer
        name:
          type: string
        posterPath:
          type: string
        posterUrl:
          type: string
        seasonCount:
          type: integer
        showKey:
          type: string
      type: object
    sudoStream_internal_dlna.Settings:
      properties:
        enabled:
          type: boolean
        udn:
          description: UDN is a stable uuid:… for SSDP (generated on first save when
            empty).
          type: string
        userId:
          type: string
      type: object
...
    sudoStream_internal_homeshelf.ShelfItem:
      properties:
        durationSeconds:
          type: number
        librarySlug:
          type: string
        path:
          type: string
        positionSeconds:
          type: number
        posterUrl:
          type: string
        title:
          type: string
      type: object
...
    sudoStream_internal_mediafs.Action:
      properties:
        canDelete:
          type: boolean
        download:
          type: string
        play:
          type: string
        thumbnail:
          type: string
      type: object
...
info:
  description: HTTP API for browsing and streaming media files.
  title: sudoStream API
  version: 0.35.0
openapi: 3.1.0
paths:
  /api/admin/dlna/settings:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sudoStream_internal_dlna.Settings'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Internal Server Error
      summary: Get DLNA settings
      tags:
      - admin
    patch:
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/sudoStream_internal_dlna.Settings'
                description: DLNA settings
                summary: body
        description: DLNA settings
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sudoStream_internal_dlna.Settings'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Internal Server Error
      summary: Update DLNA settings
      tags:
      - admin
...
  /api/catalog/{librarySlug}:
    get:
      description: Film movie cards or series show cards for a typed library (paginated).
      parameters:
      - description: Library slug
        in: path
        name: librarySlug
        required: true
        schema:
          type: string
      - description: Page size (default 28, max 100)
        in: query
        name: limit
        schema:
          type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sudoStream_internal_catalog.LibraryCatalog'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Not Found
      summary: Library catalog
      tags:
      - media
  /api/catalog/{librarySlug}/shows/{showKey}:
    get:
      description: Season/special headers with episode counts (episodes loaded separately).
      parameters:
      - description: Library slug
        in: path
        name: librarySlug
        required: true
        schema:
          type: string
      - description: Normalized show key
        in: path
        name: showKey
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sudoStream_internal_catalog.ShowDetail'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Not Found
      summary: Series show catalog
      tags:
      - media
  /api/catalog/{librarySlug}/shows/{showKey}/seasons/{season}/episodes:
    get:
      description: Episodes for one season of a show (full season when limit omitted).
      parameters:
      - description: Library slug
        in: path
        name: librarySlug
        required: true
        schema:
          type: string
      - description: Normalized show key
        in: path
        name: showKey
        required: true
        schema:
          type: string
      - description: Season number (0 = specials)
        in: path
        name: season
        required: true
        schema:
          type: integer
      - description: Optional page size (omit for full season; max 100)
        in: query
        name: limit
        schema:
          type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sudoStream_internal_catalog.SeasonEpisodes'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Not Found
      summary: Series season episodes
      tags:
      - media
...
  /api/favorite/{path}:
    get:
      description: Returns whether the current user has favorited a media file
      parameters:
      - description: Media file path
        in: path
        name: path
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.FavoriteResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Internal Server Error
      summary: Get favorite status
      tags:
      - media
    patch:
      description: Marks or clears favorite for a media file for the current user
      parameters:
      - description: Media file path
        in: path
        name: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_httpapi.FavoritePatchRequest'
                description: Favorite flag
                summary: body
        description: Favorite flag
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.FavoriteResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Internal Server Error
      summary: Update favorite status
      tags:
      - media
...
  /api/me/favorites:
    get:
      parameters:
      - description: Maximum items (default 28, max 100)
        in: query
        name: limit
        schema:
          type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.HomeItemsResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Internal Server Error
      summary: List home favorites
      tags:
      - home
...
servers:
- description: Local development server
  url: http://localhost:8080